### 2022-11-16 #4939 @dfm
@tomkimpson — Thanks for your submission! Given the relatively small scope of this submission the editorial team will take a look to decide whether or not this meets the [JOSS submission requirements](https://joss.readthedocs.io/en/latest/submitting.html). This normally takes about a week and I'll get back to you ASAP.

If there is any additional context that might help our discussion that is not included in the docs or paper, please feel free to share it here, but that is not required. Thanks!
