### 2022-11-16 #4939 @tomkimpson
Thanks @dfm . I appreciate the submission is quite concise, but I believe it fulfils the submission criteria, being research software that has been/will be used for academic papers. Additional context and detail is outlined in the paper.

Whilst the commit history is relatively short, this is on account of a previous, personal, unpublished, prototype Fortran code that uses some of the same concepts already having been written. Consequently, the construction of a Julia package took less time than it would otherwise starting from scratch.

The conciseness of the code is also an advantage - it uses Julia's composability to e.g. implement equations using Einstein summation notation or use an ODE solver package cf. working in Fortran and explicitly implementing an ODE solver or writing out full algebraic (i.e. non-tensorial) expressions.

Thanks for taking the time to review.

Cheers!
