### 2022-12-04 #4992 @editorialbot
[REVIEW]: RelativisticDynamics.jl: Relativistic Spin-Orbital 1 Dynamics in Julia
**Submitting author:** @tomkimpson (<a href="http://orcid.org/0000-0002-6542-6032">Tom Kimpson</a>)
**Repository:** https://github.com/tomkimpson/RelativisticDynamics.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.2
