### 2023-02-01 #4992 @duetosymmetry
Hi all, sorry for being extremely slow... I've had a lot of responsibilities lately and let this one slide. Please forgive me! To avoid further delays let me just drop some initial thoughts here. Again sorry that this is not as thorough as it should be, but I thought that getting started with something is better than delaying.

First, I think the name `RelativisticDynamics.jl` is much too general for what is actually happening in this package. Lots of things could be relativistic dynamics. It could be fully nonlinear numerical relativity, or it could be simple post-Newtonian. This package is specifically for solving the MPD equations on the Kerr background, not more or less.

If the goal is for the package to become more flexible in the future — for example, solving the MPD equations on other exact spacetimes, or perturbatively away from Kerr — then it must be organized in such a way that a user could swap out the background metric with ease. That is impossible right now. All of the information for a background metric — code for computing contra- and covariant metric components, Christoffels, components of Riemann — should be collected together in a single piece of data.

The structure of [metric.jl](https://github.com/tomkimpson/RelativisticDynamics.jl/blob/main/src/metric.jl) is rather messy (also, this is specific to Kerr, not any metric; and it's not just the metric, it's also connection coefficients and curvature). We have some functions whose names encode metric components, e.g. `metric_g14`. Then `covariant_metric` calls all of these. On the other hand, `christoffel` just computes all the Christoffel coefficients, without dispatching to anything like `christoffel_111`. Same thing for `riemann`. That is of course better. I don't know why there should be separate `metric_gAB` functions.

In some places in this code, there is reference to `RelativisticDynamics.delta` and `RelativisticDynamics.sigma`, while in other places it's just `delta` and `sigma`. I'm not sure why these are relegated to `useful_functions.jl`, or what determines why the functions in there ended up there. I would also recommend to use the relativists' naming to distinguish between the Levi-Civita _symbol_ (which takes values either ±1 or 0) versus the Levi-Civita _tensor_ (which transforms correctly under coordinate transformations).

Storage in `Rtensor` is extremely wasteful. There are only 10 independent quantities, but `Rtensor` allocates 256. This can be avoided be everywhere rewriting Riemann in terms of its electric/magnetic decomposition, both of these STF tensors having only 5 independent components (though needing to store 6, or more likely 9; still a huge improvement). This is not really a big point, since we're integrating an ODE, not solving a PDE. But still, I thought I should bring up these kind of design decisions.

I don't know why there is a separate `schwarzchild_covariant_riemann` function — and why does it have the spin, `a`, as an argument, when Schwarzschild is non-spinning? I would have thought that to make it obviously correct, one should just dispatch to `riemann` but with `a=0`, and lower the first index with the metric.

The documentation in `orbit.jl` refers to a file `parameters.jl` but there is no such file.

In `universal_constants.jl`, the speed of light should not be 3e8. It is defined to be exactly 299792458 m/s in SI units. In that same file: a capital G should probably be used for Newton's G. However, G is only measured to a level of 2*10^{-5}, and the mass of the Sun is similarly uncertain. However the product GM_☉ is known to a level of about 10^{-10}. It should be clear how to rewrite things so that you only ever use the product GM_☉ instead of either G or M_☉ by themselves (as is done in the pulsar timing and GR communities).

There are really many more issues about the clarity and organization of the code along these same lines, so I will not try to enumerate any more. In general, I think the code should organized better.

Turning to the documentation and article: I think the issue of spin supplementary conditions needs to be discussed much, much more. The SSCs are the main point of confusion when it comes to the MPD equations. Some statements you make in the code / documentation are only true for one particular SSC. It's fine to stick to one, as long as you make it clear that no others are allowed, and point out which statements are specific to the SSC you chose. It would be helpful to point to a specialist article about different SSCs.

Going back to the big picture: If the point of a specialized package like this one is for spinning bodies just in the Kerr spacetime, and you want to have high precision over very long times, then it would be better to use an action-angle formulation (see e.g. Vojtech Witzany's papers). That is only appropriate for the first order in spin, but as the article rightly mentions, the multipole expansion of the small body has already been truncated to dipole order — so the MPD equations are already an expansion. Of course the AA formulation is specific to Kerr. It would still be relevant spacetimes close to Kerr, perturbation theory. On the other hand to make it totally general, for any background spacetime, you can't use AA variables. I think it would be appropriate, in the Statement of Need, to discuss why this has been formulated as a generic ODE integration problem, rather than specializing to AA variables.

Sorry again that I waited so long to say anything. I hope this gives some useful directions for improvement of both the code and documentation. We can continue to discuss.
