### 2023-02-06 #4992 @tomkimpson
Thanks @duetosymmetry for the thorough comments. No worries re the delay - I understand you're busy and appreciate you taking the time to review. I'll open a PR and start working through your comments and ping for any further discussion. Cheers!
