### 2023-02-22 dc7277c
check number format is conserved
