+++ test/runtests.jl
@@ -12,4 +12,5 @@ include("initial_conditions.jl")
 include("universal_constants.jl")
 include("initial_conditions.jl")
 include("timestepping.jl")
+include("number_formats.jl")
 include("orbit.jl")
