### 2023-02-24 10217ff
overhaul metric.jl, reduce useful_functions.jl
