+++ src/orbit.jl
@@ -39,4 +39,12 @@ function bounds_checks(P::SystemParameters)
 end 
 
 
+"""
+    M = Model(P,C) 
+The model struct which holds all the parameters (P) and constants (C)
+"""
+struct Model{NF<:AbstractFloat} <: ModelSetup
+    parameters::SystemParameters
+    constants::Constants{NF}
+end
 
