+++ test/metric.jl
@@ -1,13 +1,13 @@
 using RelativisticDynamics
 using Test
-using Zygote
+using Enzyme
 using TensorOperations
 using LinearAlgebra
 using Distributions
 
 @testset "Trace of the metric" begin
     
-    for n in 1:3
+    for n in 1:5
 
         #Get some coordiantes at random 
         r = rand(Uniform(3.0,1e5))      # Radial coordinate. 3.0 as rough lower limit of an event horizon
@@ -32,7 +32,7 @@ end
 @testset "Christoffel tensor components" begin
     
 
-    for n in 1:3
+    for n in 1:5
 
         #Get some coordiantes at random 
         t = rand(Uniform(3.0,1e5))       # Time coordinate - arbitratry since metric is time-independent
@@ -41,30 +41,26 @@ end
         ϕ = rand(Uniform(0.0, 2.0*π))    # Azimuth coordinate - arbitratry since metric is axisymmetric
         a = rand(Uniform(-0.99, 0.99))   # Spin parameter
 
-        coords = [t,r,θ,ϕ]
+        x = [t r θ ϕ]
         
         #Calculate the metric 
-        g = RelativisticDynamics.covariant_metric(coords,a)
-        g_inverse = RelativisticDynamics.contravariant_metric(coords,a)
+        g = RelativisticDynamics.covariant_metric(x,a)
+        g_inverse = RelativisticDynamics.contravariant_metric(x,a)
 
 
-        #Use automatic diff to get the first derivatives of the metric  of the metric 
-        #Only calculate r and θ derivatives
-        #`jacobian()` returns a tuple so [1] extracts the vector of floats 
-        g_∂ = zeros(Float64,4,4,4) # a derivative tensor 
-
-        g_∂[1,1,:] = jacobian(x -> RelativisticDynamics.metric_g11(x,a), [t r θ ϕ])[1]
-        g_∂[2,2,:] = jacobian(x -> RelativisticDynamics.metric_g22(x,a), [t r θ ϕ])[1]
-        g_∂[3,3,:] = jacobian(x -> RelativisticDynamics.metric_g33(x,a), [t r θ ϕ])[1]
-        g_∂[4,4,:] = jacobian(x -> RelativisticDynamics.metric_g44(x,a), [t r θ ϕ])[1]
-        g_∂[1,4,:] = jacobian(x -> RelativisticDynamics.metric_g14(x,a), [t r θ ϕ])[1]
-
+        #Use automatic diff to get the first derivatives of the metric
+        g_∂ = zeros(Float64,4,4,4) # a derivative tensor
+        shadow = onehot(x)
+        for i = 1:length(shadow)
+            g_∂[:,:,i] = autodiff(Forward, RelativisticDynamics.covariant_metric, DuplicatedNoNeed, Duplicated(x, shadow[i]), Const(a))[1]
+        end 
+        
         @tensor begin
             Γ[μ,ν,λ] := 0.50*g_inverse[μ,ρ]*(g_∂[ρ,ν,λ] +g_∂[ρ,λ,ν]-g_∂[ν,λ,ρ])  #:= allocates a new array
         end
 
         #Compare with the analytical solution
-        Γ_analytical = RelativisticDynamics.christoffel(coords,a)
[…]
