+++ test/universal_constants.jl
@@ -1,8 +1,8 @@
 using RelativisticDynamics
 using Test
-using Zygote
+#using Zygote
 using TensorOperations
-using LinearAlgebra
+#using LinearAlgebra
 using Distributions
 
 
