+++ src/initial_conditions.jl
@@ -13,7 +13,7 @@ end
 """
     initialization = initial_conditions(M)
 Setup the initial conditions for the MPD orbital dynamics"""
-function initial_conditions(M::Model)
+function initial_conditions(M)
 
 
     @unpack NF = M.parameters
