+++ Project.toml
@@ -4,52 +4,12 @@ authors = ["Tom Kimpson"]
 version = "0.1.0"
 
 [deps]
-ChainRulesCore = "d360d2e6-b24c-11e9-a2a3-2a2ae2dbcce4"
 Combinatorics = "861a8166-3701-5b0c-9a16-15d98fcdc6aa"
-ComponentArrays = "b0b7db55-cfe3-40fc-9ded-d10e2dbeff66"
 DifferentialEquations = "0c46a032-eb83-5123-abaf-570d42b7fbaa"
-Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
-Einsum = "b7d42ee7-0b51-5a75-98ca-779d3107e4c0"
-Enzyme = "7da242da-08ed-463a-9acd-ee780be4f1d9"
-ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
-JLD = "4138dd39-2aa7-5051-a626-17a0bb65d9c8"
 LaTeXStrings = "b964fa9f-0449-5b57-a5c2-d3ea65f4040f"
-LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
-PlotlyJS = "f0f68f2c-4968-5e81-91da-67840de0976a"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
-PreallocationTools = "d236fae5-4411-538c-8e31-a6e3d9e00b46"
-Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
-ProtoStructs = "437b6fc4-8e8e-11e9-3fa1-ad391e66c018"
-Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
-SciMLSensitivity = "1ed8b502-d754-442c-8d5d-10ac956f44a1"
-TensorOperations = "6aa20fa7-93e2-5fca-9bc0-fbd0db3c71a2"
 Tullio = "bc48ee85-29a4-5162-ae0b-a64e1601d4bc"
-UnicodePlots = "b8865327-cd53-5732-bb35-84acbb429228"
-Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
-
-[compat]
-ChainRulesCore = "1"
-Combinatorics = "1"
-ComponentArrays = "0.13"
-DifferentialEquations = "7"
-Distributions = "0.25"
-Einsum = "0.4"
-ForwardDiff = "0.10"
-JLD = "0.13"
-LaTeXStrings = "1"
-Parameters = "0.12"
-PlotlyJS = "0.18"
-Plots = "1"
-PreallocationTools = "0.4"
-ProtoStructs = "1"
-Revise = "3"
-SciMLSensitivity = "7"
-TensorOperations = "3"
-Tullio = "0.3"
-UnicodePlots = "3"
-Zygote = "0.6"
-julia = "1.1"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
