+++ paper/paper.md
@@ -46,10 +46,11 @@ which when expanded into a set of infinite multipole moments leads to a descript
 \begin{equation}\label{eq:mpd2}
 \frac{Ds^{\mu \nu}}{d \lambda} =p^{\mu}u^{\nu} - p^{\nu}u^{\mu}
 \end{equation}
-for affine parameter $\lambda$, 4-velocity $u^{\nu}$ and Riemann curvature tensor ${R^{\mu}}_{\nu \alpha \beta}$.  The system is closed by providing a spin supplementary condition
+for affine parameter $\lambda$, 4-velocity $u^{\nu}$ and Riemann curvature tensor ${R^{\mu}}_{\nu \alpha \beta}$.  The system is closed by providing a spin supplementary condition, equivalent to specifying the observer-dependent  centre of mass. For this work we take the Tulczyjew-Dixon condition [@tulczyjew; @Dixon1964]
 \begin{equation}\label{eq:mpd3}
 s^{\mu \nu} p_{\nu} = 0
 \end{equation}
+
 Together, equations \ref{eq:mpd1} - \ref{eq:mpd3} form the Mathisson-Papetrou-Dixon (MPD) equations [@Mathisson1937;@Papapetrou1951; @Dixon1964], and describe the spin-orbital evolution in a fully consistent way that is applicable to strong field regimes. 
 
 
@@ -58,7 +59,7 @@ Together, equations \ref{eq:mpd1} - \ref{eq:mpd3} form the Mathisson-Papetrou-Di
 
 `RelativisticDynamics.jl` is an open-source Julia package for relativistic spin-orbital dynamics in the gravitational strong field for a Kerr spacetime. Existing codes for modelling the dynamics of spinning objects like pulsars in the strong-field regime are generally lacking, since such systems occupy an intermediate regime that is generally overlooked. At the "low" end of this regime there are post-Newtonian or geodesic descriptions which neglect the influence of the pulsar spin on the underlying spacetime metric ("spin-curvature" coupling). At the "high" end there is the full Numerical Relativity (NR) solutions which are primarily applicable to two BHs with a mass ratio $\mathcal{O}(1)$, and are computationally intractable for these MSP systems which are observed over a large number of orbital cycles.  
 
-`RelativisticDynamics.jl` aims to bridge this gap by providing a modern, fast code for accurate numerical evolution of spinning relativistic systems, via the MPD formalism. Julia is a modern language that solves the "two language problem", enabling fast dynamic typing and JIT compilation in conjunction with petaflop performance, comparable with numerical languages that are better known in the astrophysics community such as C or Fortran. As a modern language, it also provides a dedicated package manager and a large catalogue of _composable_ packages for scientific computing. This enables `RelativisticDynamics.jl` to easily leverage and interface with other scientific computing packages. The author and collaborators have used the general methods and mathematics described in this package for multiple research projects [e.g. @Kimpson2019; @Li2019; @Kimpson2020; @KimpsonAA] with a particular focus on the radio signals from spinning pulsar systems.  This package represents an attempt to create a documented, well-tested, open source resource for public use in this area, that can also be used as a computational playground for exploring techniques that could be applicable to more advanced numerical models.
[…]
