+++ paper/paper.bib
@@ -140,3 +140,31 @@ archivePrefix = {arXiv},
       pages=      "495 - 502",
       url = "https://journals.ametsoc.org/view/journals/mwre/145/2/mwr-d-16-0228.1.xml"
 }
+
+@article {tulczyjew,
+      author = "Tulczyjew, W. M.",
+      title = "Motion of multipole particles in general relativity theory binaries",
+      journal = "Acta Phys. Pol.",
+      year = "1959",
+      volume = "18",
+      pages=      "393",
+}
+
+
+@ARTICLE{Witzany2022,
+       author = {{Witzany}, Vojt{\v{e}}ch},
+        title = "{Action-angle coordinates for black-hole geodesics I: Spherically symmetric and Schwarzschild}",
+      journal = {arXiv e-prints},
+     keywords = {General Relativity and Quantum Cosmology, Astrophysics - High Energy Astrophysical Phenomena},
+         year = 2022,
+        month = mar,
+          eid = {arXiv:2203.11952},
+        pages = {arXiv:2203.11952},
+          doi = {10.48550/arXiv.2203.11952},
+archivePrefix = {arXiv},
+       eprint = {2203.11952},
+ primaryClass = {gr-qc},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2022arXiv220311952W},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
