+++ test/universal_constants.jl
@@ -1,7 +1,6 @@
 using RelativisticDynamics
 using Test
 #using Zygote
-using TensorOperations
 #using LinearAlgebra
 using Distributions
 
@@ -11,32 +10,6 @@ using Combinatorics
 
 
 
-@testset "Zero spin MPD case for mapping functions" begin
-    
-    NF = Float64
-
-    m = :MPD
-    a = 0.0
-
-    for i in 1:5
-        r = rand(Uniform(3.0,1e5)) 
-        θ = rand(Uniform(0.0, 2.0*π))
-        
-        #Mapping functions
-        fr = RelativisticDynamics.mapping_f(r,a,cos(θ))
-        gr = RelativisticDynamics.mapping_g(r,a)
-        hr = RelativisticDynamics.mapping_h(r,a,cos(θ))
-        dr = RelativisticDynamics.mapping_d(r,a,cos(θ))
-
-        #Zero spin definitions
-        @test fr == r^4
-        @test gr == 0.0
-        @test isapprox(hr, r*(r-2.0) + (cos(θ)^2)/(1.0 - cos(θ)^2) *(r^2-2.0*r))
-        @test isapprox(dr, (r^2)*(r^2-2.0*r))
-
-    end 
-    
-end
 
 
 @testset "MPD Carter constant for zero inclination" begin
