+++ test/number_formats.jl
@@ -1,12 +1,4 @@
 
-
-
-using RelativisticDynamics
-using Test
-import Parameters: @with_kw, @unpack
-using Tullio
-
-
 @testset "Check all number formats are consistent" begin
 
     for NF in [Float64,Float32]
