+++ src/plotting.jl
@@ -3,7 +3,6 @@
 
 using Plots 
 using LaTeXStrings
-#using Printf
 
 """
     PlotTrajectory(solution,model,dimensions=[1,2,3],savepath="")
@@ -15,8 +14,8 @@ function PlotTrajectory(solution,model,dimensions=[1,2,3],savepath="")
 
     @unpack a = model.parameters    #Get the BH spin parameter 
     
-    println("Plotting the solution generated with the following user defined parameters")
-    display(model.parameters)
+    println("Plotting the solution generated with the below user-defined parameters")
+    println(model.parameters)
     println("-------------------------------")
 
     #Interpolate to higher resolution for smooth plotting   
