+++ src/orbit.jl
@@ -8,6 +8,8 @@ function orbit(::Type{NF}=Float64;              # number format, use Float64 as 
                kwargs...                        # all additional non-default parameters
                ) where {NF<:AbstractFloat}
 
+
+    
     # Setup all system parameters, universal constants etc.
     P = SystemParameters(NF=NF;kwargs...) # Parameters
     bounds_checks(P)                      # Check all parameters are reasonable
@@ -16,9 +18,10 @@ function orbit(::Type{NF}=Float64;              # number format, use Float64 as 
 
     #Initial conditions 
     initialization = initial_conditions(M)
-
+    
     #Evolve in time
     solution = timestepping(initialization, M)
+    
     return solution, M
 
 end
