### 2023-04-24 #4992 @duetosymmetry
Ping @farr for any review.

Question for @dfm: Should we be reviewing the PR tomkimpson/RelativisticDynamics.jl#41, or should @tomkimpson merge it in if he sees fit, and we do another round of refereeing back here? (Tangentially related, are the Julia docs also being generated from the PR branch? I can't find those separately).
