### 2023-04-24 #4992 @dfm
Thanks for checking in @tomkimpson and @duetosymmetry! I'll email @farr to remind him as well.

> Should we be reviewing the PR https://github.com/tomkimpson/RelativisticDynamics.jl/pull/41, or should @tomkimpson merge it in if he sees fit, and we do another round of refereeing back here? (Tangentially related, are the Julia docs also being generated from the PR branch? I can't find those separately).

Good question! These reviews can progress either way. It's actually recommended to do as much of the iteration as you can on PRs and issues in the parent repository, rather than directly in this thread. So it certainly could be a good approach to go through that PR directly, but we're flexible, so please use whichever approach works best for you!
