### 2023-05-24 #4992 @tomkimpson
Thanks @dfm. All of @duetosymmetry comments were address in https://github.com/tomkimpson/RelativisticDynamics.jl/pull/41 . I have been leaving this PR open for more comments, but if @duetosymmetry is happy I will close it while we wait for the new reviewer.

One question for @dfm : it was recommended to change the package name to be less general. This is obviously straightforward but are there any issue on the JOSS side if I update the github repo name?
