### 2023-05-30 #4992 @dfm
> One question for @dfm : it was recommended to change the package name to be less general. This is obviously straightforward but are there any issue on the JOSS side if I update the github repo name?

@tomkimpson — I'm so sorry for the delayed response! No - there is no problem with this. Just ping me once you've changed the name!
