### 2023-05-31 #4992 @langfzac
Hi all, I have some initial comments after reading through the paper/docs, and trying out the example code.

Paper:
- The paper mentions that this code is meant to fill the space between full NR and simple post-Newtonian treatments. It would be good to include references to existing codes.

Documentation:
- "How to run RelativisticDynamics.jl" section:
  - The user is pointed to a file `src/default_parameters.jl`, which (I assume) should be `src/system_parameters.jl`.
  - What is the `run_speedy()` function? I don't see it defined anywhere in the repo.
- "Visualisation of the solution":
  - The 3D `PlotTrajectory()` throws an error: ` UndefVarError: 'plt' not defined`.
  - `StackedPlot` throws an error: ` type Model has no field constant`

Example Notebook:
- Supplying the `NF=Float32` keyword to `orbit()` throws an error: `No matching function wrapper was found!` on Julia v1.9. However, it does work on v1.7.
- I think this is already known, but autodiff appears to be broken. It looks like in the `dts_comments_1` branch, the notebook has an updated autodiff section, which is marked as "TODO".

Tests/Continuous Integration:
- I think the tests should be running on at least the Long-Term-Support version of Julia. Preferably, both LTS and the latest release. Currently, they only run on v1.7.
- Since the ability to use autodiff is a highlighted feature, there should be a relevant unit test for `orbit()`.
- Consider using a [doctest](https://documenter.juliadocs.org/stable/man/doctests/) for the "how to run" example, and perhaps the example notebook. This would catch any issues with the example code, and make sure it stays up-to-date.
- Some of the test (e.g. `test/orbit.jl`) check for things like misspecified parameters via try-catch, but only check that *any* error is thrown. These could be made more explicit by testing for the particular error that is expected.

General:
- It might be worth reworking the plotting functions in the [Recipes](https://docs.juliaplots.org/stable/recipes/) framework from `Plots.jl`. It would remove the `Plots` dependency (which is historically a fairly heavy-weight one), and allow more flexibility for the user. Also, `PlotTrajectory()` does not display in Pluto notebooks, which would likely be fixed.
- I think there are some dependencies that could be dropped. For example, it doesn't look like `Zygote` or `SciMLSensitivity` are explicitly being used in any functions within the module. Others are `BenchmarkTools`, `Enzyme`, and `ForwardDiff`. It looks like `Distributions` is only used within the testing suite, which can have it's own `Project.toml` (like the docs).

Hopefully this is helpful! @tomkimpson please let me know if you need clarification on anything.
