+++ paper/paper.md
@@ -57,7 +57,7 @@ Together, equations \ref{eq:mpd1} - \ref{eq:mpd3} form the Mathisson-Papetrou-Di
 <!-- A Statement of need section that clearly illustrates the research purpose of the software and places it in the context of related work -->
 # Statement of need
 
-`RelativisticDynamics.jl` is an open-source Julia package for relativistic spin-orbital dynamics in the gravitational strong field for a Kerr spacetime. Existing codes for modelling the dynamics of spinning objects like pulsars in the strong-field regime are generally lacking, since such systems occupy an intermediate regime that is generally overlooked. At the "low" end of this regime there are post-Newtonian or geodesic descriptions which neglect the influence of the pulsar spin on the underlying spacetime metric ("spin-curvature" coupling). At the "high" end there is the full Numerical Relativity (NR) solutions which are primarily applicable to two BHs with a mass ratio $\mathcal{O}(1)$, and are computationally intractable for these MSP systems which are observed over a large number of orbital cycles.  
+`RelativisticDynamics.jl` is an open-source Julia package for relativistic spin-orbital dynamics in the gravitational strong field for a Kerr spacetime. Existing codes for modelling the dynamics of spinning objects like pulsars in the strong-field regime are generally lacking, since such systems occupy an intermediate regime that is generally overlooked. At the "low" end of this regime there are post-Newtonian or geodesic descriptions which neglect the influence of the pulsar spin on the underlying spacetime metric ("spin-curvature" coupling). At the "high" end there is the full Numerical Relativity (NR) solutions [e.g. @Andrade2021] which are primarily applicable to two BHs with a mass ratio $\mathcal{O}(1)$, and are computationally intractable for these MSP systems which are observed over a large number of orbital cycles.  
 
 `RelativisticDynamics.jl` aims to bridge this gap by providing a modern, fast code for accurate numerical evolution of spinning relativistic systems, via the MPD formalism. Julia is a modern language that solves the "two language problem", enabling fast dynamic typing and JIT compilation in conjunction with petaflop performance, comparable with numerical languages that are better known in the astrophysics community such as C or Fortran. As a modern language, it also provides a dedicated package manager and a large catalogue of _composable_ packages for scientific computing. This enables `RelativisticDynamics.jl` to easily leverage and interface with other scientific computing packages. The author and collaborators have used the general methods and mathematics described in this package for multiple research projects [e.g. @Kimpson2019; @Li2019; @Kimpson2020; @KimpsonAA] with a particular focus on the radio signals from spinning pulsar systems.  This package represents an attempt to create a documented, well-tested, open source resource for public use in this area, that can also be used as a computational playground for exploring techniques that could be applicable to more advanced numerical models. The package has been formulated in terms of ODE integration, rather than using e.g. action-angle variables [@Witzany2022], to allow for extension to general spacetime metrics and straightforward computatin of quantities relevant for pulsar observations e.g. spin axis orientation. 
 
