+++ paper/paper.bib
@@ -168,3 +168,14 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@article{Andrade2021, 
+         doi = {10.21105/joss.03703}, 
+         url = {https://doi.org/10.21105/joss.03703}, 
+         year = {2021}, 
+         publisher = {The Open Journal}, 
+         volume = {6}, 
+         number = {68}, 
+         pages = {3703}, 
+         author = {Tomas Andrade and Llibert Areste Salo and Josu C. Aurrekoetxea and Jamie Bamber and Katy Clough and Robin Croft and Eloy de Jong and Amelia Drew and Alejandro Duran and Pedro G. Ferreira and Pau Figueras and Hal Finkel and Tiago Fran\c{c}a and Bo-Xuan Ge and Chenxia Gu and Thomas Helfer and Juha Jäykkä and Cristian Joana and Markus Kunesch and Kacper Kornet and Eugene A. Lim and Francesco Muia and Zainab Nazari and Miren Radia and Justin Ripley and Paul Shellard and Ulrich Sperhake and Dina Traykova and Saran Tunyasuvunakool and Zipeng Wang and James Y. Widdicombe and Kaze Wong}, 
+         title = {GRChombo: An adaptable numerical relativity code for fundamental physics}, 
+         journal = {Journal of Open Source Software} }
