### 2023-06-04 #4992 @tomkimpson
Thanks @langfzac for the very helpful comments. I have opened a PR https://github.com/tomkimpson/RelativisticDynamics.jl/pull/42  and will ping you once all comments have been addressed
