### 2023-06-05 189d398
CompatHelper: add new compat entry for Zygote at version 0.6, (keep existing compat)
