### 2023-06-05 da50787
CompatHelper: add new compat entry for Tullio at version 0.3, (keep existing compat)
