### 2023-06-05 6589206
CompatHelper: add new compat entry for LaTeXStrings at version 1, (keep existing compat)
