### 2023-06-05 4e83efc
CompatHelper: add new compat entry for DifferentialEquations at version 7, (keep existing compat)
