### 2023-06-06 dec8bf2
remove plotting functions, just use recipe
