+++ paper/paper.bib
@@ -179,3 +179,20 @@ archivePrefix = {arXiv},
          author = {Tomas Andrade and Llibert Areste Salo and Josu C. Aurrekoetxea and Jamie Bamber and Katy Clough and Robin Croft and Eloy de Jong and Amelia Drew and Alejandro Duran and Pedro G. Ferreira and Pau Figueras and Hal Finkel and Tiago Fran\c{c}a and Bo-Xuan Ge and Chenxia Gu and Thomas Helfer and Juha Jäykkä and Cristian Joana and Markus Kunesch and Kacper Kornet and Eugene A. Lim and Francesco Muia and Zainab Nazari and Miren Radia and Justin Ripley and Paul Shellard and Ulrich Sperhake and Dina Traykova and Saran Tunyasuvunakool and Zipeng Wang and James Y. Widdicombe and Kaze Wong}, 
          title = {GRChombo: An adaptable numerical relativity code for fundamental physics}, 
          journal = {Journal of Open Source Software} }
+
+
+
+@article{PhysRevD.45.1840,
+  title = {Strong-field tests of relativistic gravity and binary pulsars},
+  author = {Damour, Thibault and Taylor, J. H.},
+  journal = {Phys. Rev. D},
+  volume = {45},
+  issue = {6},
+  pages = {1840--1868},
+  numpages = {0},
+  year = {1992},
+  month = {Mar},
+  publisher = {American Physical Society},
+  doi = {10.1103/PhysRevD.45.1840},
+  url = {https://link.aps.org/doi/10.1103/PhysRevD.45.1840}
+}
