+++ src/metric.jl
@@ -1,5 +1,129 @@
-#using SciMLSensitivity
-using Zygote
+
+# Define the metric in terms of pure functions 
+# This is more verbose and explicit than the "normal" way of defining a metric 
+# but enables AD avoiding mutation by using Zygote https://fluxml.ai/Zygote.jl/stable/limitations/
+
+
+# """
+#     g = metric_g11(coords,a)
+# The covariant tt component of the Kerr metric in Boyer Lindquist coordinates
+# """
+# function metric_g11(x,a)
+#     t,r,θ,ϕ =  x[1],x[2],x[3],x[4]
+#     Σ = RelativisticDynamics.sigma(r,θ,a)
+#     return -(1.0 - 2.0*r / Σ)
+# end 
+
+
+# """
+#     g = metric_g22(coords,a)
+# The covariant rr component of the Kerr metric in Boyer Lindquist coordinates
+# """
+# function metric_g22(x,a)
+#     t,r,θ,ϕ =  x[1],x[2],x[3],x[4]
+#     Σ = RelativisticDynamics.sigma(r,θ,a)
+#     Δ = RelativisticDynamics.delta(r,a)
+#     return Σ / Δ
+# end 
+
+# """
+#     g = metric_g33(coords,a)
+# The covariant θθ component of the Kerr metric in Boyer Lindquist coordinates
+# """
+# function metric_g33(x,a)
+#     t,r,θ,ϕ =  x[1],x[2],x[3],x[4]
+#     Σ = RelativisticDynamics.sigma(r,θ,a)
+#     return Σ 
+# end 
+
+# """
+#     g = metric_g44(coords,a)
+# The covariant ϕϕ component of the Kerr metric in Boyer Lindquist coordinates
+# """
+# function metric_g44(x,a)
+#     t,r,θ,ϕ =  x[1],x[2],x[3],x[4]
+#     Σ = RelativisticDynamics.sigma(r,θ,a)
+#     Δ = RelativisticDynamics.delta(r,a)
+#     return sin(θ)^2 * ((r^2 +a^2)^2 - Δ*a^2*sin(θ)^2) / Σ
+# end 
+
+# """
+#     g = metric_g14(coords,a)
+# The covariant tϕ component of the Kerr metric in Boyer Lindquist coordinates
+# """
+# function metric_g14(x,a)
+#     t,r,θ,ϕ =  x[1],x[2],x[3],x[4]
+#     Σ = RelativisticDynamics.sigma(r,θ,a)
+#     return -2.0*a*r*sin(θ)^2/Σ
+# end 
+
+
+# """
+#     g = metric_g11(coords,a)
+# The covariant tt component of the Kerr metric in Boyer Lindquist coordinates
+# """
+# function metric_contra_g11(x,a)
+#     t,r,θ,ϕ =  x[1],x[2],x[3],x[4]
+#     Σ = RelativisticDynamics.sigma(r,θ,a)
+#     Δ = RelativisticDynamics.delta(r,a)
+#     covar = sin(θ)^2 * ((r^2 +a^2)^2 - Δ*a^2*sin(θ)^2) / Σ
+#     denom = Δ*sin(θ)^2
+#     return -covar/denom 
+# end 
+
+
+# """
+#     g = metric_g22(coords,a)
+# The covariant rr component of the Kerr metric in Boyer Lindquist coordinates
+# """
+# function metric_contra_g22(x,a)
+#     t,r,θ,ϕ =  x[1],x[2],x[3],x[4]
+#     Σ = RelativisticDynamics.sigma(r,θ,a)
+#     Δ = RelativisticDynamics.delta(r,a)
[…]
