### 2023-06-12 83c0db5
unit tests for metric.jl
