+++ test/orbit.jl
@@ -15,8 +15,8 @@ end
     try
         solution,model = RelativisticDynamics.orbit(a=1.1)
         @test false 
-    catch e
-        @test  true
+    catch e 
+        @test e == ErrorException("Spin parameter a is out of bounds")
     end
 
     #Relative masses not correct
@@ -24,7 +24,7 @@ end
         solution,model = RelativisticDynamics.orbit(mBH =1e6, mPSR=1e6)
         @test false  
     catch e
-        @test  true
+        @test e == ErrorException("Mass ratio is too small")
     end
 
 
@@ -33,7 +33,7 @@ end
         solution,model = RelativisticDynamics.orbit(rPSR=1e6)
         @test false  
     catch e
-        @test  true
+        @test e == ErrorException("Pulsar radius is unphysical")
     end
 
 
@@ -42,24 +42,31 @@ end
         solution,model = RelativisticDynamics.orbit(e=1.9)
         @test false  
     catch e
-        @test  true
+        @test e == ErrorException("Eccentricity is outside range")
     end
 
-    #Orbit dir
+    #Orbit dir must be 1/-1
+    try
+        solution,model = RelativisticDynamics.orbit(orbit_dir=2)
+        @test false  
+    catch e
+        @test e == ErrorException("Orbit direction must be plus or minus 1")
+    end
+
+    #Orbit dir must be integer
     try
         solution,model = RelativisticDynamics.orbit(orbit_dir=1.1)
         @test false  
     catch e
-        @test  true
-    end
-
+        @test e == InexactError(:Int64, Int64, 1.1)
+    end 
 
     #Incliation
     try
         solution,model = RelativisticDynamics.orbit(ι=10.0)
         @test false  
     catch e
-        @test  true
+        @test e == ErrorException("ι is outside of allowed range")
     end
 
 end
