+++ test/metric.jl
@@ -1,4 +1,7 @@
-
+using Tullio
+using Distributions
+using LinearAlgebra
+using Zygote
 @testset "Trace of the metric" begin
     
     for n in 1:5
@@ -12,10 +15,9 @@
         coords = [0.0,r,θ,0.0]
         g = RelativisticDynamics.covariant_metric(coords,a)
         g_inverse = RelativisticDynamics.contravariant_metric(coords,a)
-        #Check metric trace is OK 
         
+        #Check metric trace is OK 
         @tullio δ[a,c] := g[a,b] * g_inverse[b,c]  #:= allocates a new array
-        #end
         @test isapprox(tr(δ),4.0)
     end
 
@@ -35,211 +37,77 @@ end
         ϕ = rand(Uniform(0.0, 2.0*π))    # Azimuth coordinate - arbitratry since metric is axisymmetric
         a = rand(Uniform(-0.99, 0.99))   # Spin parameter
 
-        x = [t r θ ϕ]
+        coords = [t,r,θ,ϕ]
         
         #Calculate the metric 
-        g = RelativisticDynamics.covariant_metric(x,a)
-        g_inverse = RelativisticDynamics.contravariant_metric(x,a)
+        g = RelativisticDynamics.covariant_metric(coords,a)
+        g_inverse = RelativisticDynamics.contravariant_metric(coords,a)
 
+        #and the metric Jacobian using Zygote
+        g_jacobian = jacobian(x -> RelativisticDynamics.covariant_metric_zygote(x,a), [t r θ ϕ])[1]
+        g_jacobian = reshape(g_jacobian,(4,4,4)) # reshape
 
-        #Use automatic diff to get the first derivatives of the metric
-        g_∂ = zeros(Float64,4,4,4) # a derivative tensor
-        shadow = onehot(x)
-        for i = 1:length(shadow)
-            g_∂[:,:,i] = autodiff(Forward, RelativisticDynamics.covariant_metric, DuplicatedNoNeed, Duplicated(x, shadow[i]), Const(a))[1]
-        end 
-        
-        @tullio Γ[μ,ν,λ] := 0.50*g_inverse[μ,ρ]*(g_∂[ρ,ν,λ] +g_∂[ρ,λ,ν]-g_∂[ν,λ,ρ])  #:= allocates a new array
-        #end
+        #Christoffel
+        @tullio Γ[μ,ν,λ] := 0.50*g_inverse[μ,ρ]*(g_jacobian[ρ,ν,λ] +g_jacobian[ρ,λ,ν]-g_jacobian[ν,λ,ρ])  #:= allocates a new array
 
         #Compare with the analytical solution
-        Γ_analytical = RelativisticDynamics.christoffel(x,a)
+        Γ_analytical = RelativisticDynamics.christoffel(coords,a)
         @test isapprox(Γ,Γ_analytical)
-
-
     end
-
 end 
-
-
-# @testset "Riemann tensor components" begin
-    
-
-#     #for n in 1:5
-#     for n in 1:1
-
-#         #Get some coordiantes at random 
-#         t = rand(Uniform(3.0,1e5))       # Time coordinate - arbitratry since metric is time-independent
[…]
