### 2023-06-13 7e510ff
basic doctest example
