+++ test/orbit.jl
@@ -1,3 +1,7 @@
+
+using Zygote
+using SciMLSensitivity
+
 @testset "Basic run through" begin
     try
         solution,model = RelativisticDynamics.orbit()
@@ -68,5 +72,35 @@ end
     catch e
         @test e == ErrorException("ι is outside of allowed range")
     end
+
+end
+
+
+@testset "Automatic differentiation with Zygote" begin
+
+    #See also notebooks/demo.ipynb
+    function loss_function(reference,e)
+
+        # Run the model using the new value of e
+        solution,model = orbit(e=e,Norbits=1)
+        
+        r0 = last(reference[2,:])  # The last r value of the reference solution
+        r = last(solution[2,:])    # The last r value of the new solution
+        loss = abs(r0 - r)         # Loss
+      
+        return loss
+    
+    end
+
+
+    reference_solution,reference_model = orbit(Norbits=1); #using the default parameters
+
+    loss_g_e01= Zygote.gradient(x -> loss_function(reference_solution,x),0.1)[1]
+    loss_g_e05 = Zygote.gradient(x -> loss_function(reference_solution,x),0.5)[1]
+    loss_g_e09 = Zygote.gradient(x -> loss_function(reference_solution,x),0.9)[1]
+ 
+    @test loss_g_e05 == 0.0 
+    @test loss_g_e01 < loss_g_e05 < loss_g_e09
+
 
 end
