### 2023-06-13 1f04c5a
clean up notebooks, Project.toml test packages
