### 2023-06-13 #4992 @tomkimpson
@langfzac Thanks for all your comments, they were very useful! I have addressed each of them in turn here: https://github.com/tomkimpson/RelativisticDynamics.jl/pull/42#issuecomment-1588472777
