+++ test/metric.jl
@@ -44,7 +44,7 @@ end
         g_inverse = RelativisticDynamics.contravariant_metric(coords,a)
 
         #and the metric Jacobian using Zygote
-        g_jacobian = jacobian(x -> RelativisticDynamics.covariant_metric_zygote(x,a), [t r θ ϕ])[1]
+        g_jacobian = jacobian(x -> RelativisticDynamics.covariant_metric(x,a), [t r θ ϕ])[1]
         g_jacobian = reshape(g_jacobian,(4,4,4)) # reshape
 
         #Christoffel
@@ -196,227 +196,6 @@ end
     end
 
 end 
-
-
-
-
-
-
-
-
-
-#Riemann tensor components test: TBD
-#Passes when using separate pure functions for each non-zero metric component, but need a method to do this for hessians applied to general buffered arrays.
-#Currently exploring the best way to do this. For Kerr not a problem since the Riemann function is well-tested, but for general metrics 
-#it will be nice to get this working properly.
-
-
-# @testset "Riemann tensor components" begin
-    
-
-#     #for n in 1:5
-#     for n in 1:1
-
-#         #Get some coordiantes at random 
-#         t = rand(Uniform(3.0,1e5))       # Time coordinate - arbitratry since metric is time-independent
-#         r = rand(Uniform(3.0,1e3))       # Radial coordinate. 3.0 as rough lower limit of an event horizon
-#         θ = rand(Uniform(0.0, 2.0*π))    # Polar coordinate
-#         ϕ = rand(Uniform(0.0, 2.0*π))    # Azimuth coordinate - arbitratry since metric is axisymmetric
-#         a = rand(Uniform(-0.99, 0.99))   # Spin parameter
-
-#         coords = [t,r,θ,ϕ]
-
-#         #Calculate the metric 
-#         println("metric")
-#         g = RelativisticDynamics.covariant_metric(coords,a)
-#         g_inverse = RelativisticDynamics.contravariant_metric(coords,a)
-
-#         # First derivatives of the covariant metric 
-#         println("metric derivative")
-#         g_jacobian = jacobian(x -> RelativisticDynamics.covariant_metric_zygote(x,a), [t r θ ϕ])[1]
-#         g_jacobian = reshape(g_jacobian,(4,4,4)) # reshape
-
-
-#         # First derivatives of the contravariant metric 
-#         println("inverse metric derivatieve")
-#         g_inverse_jacobian = jacobian(x -> RelativisticDynamics.contravariant_metric(x,a), [t r θ ϕ])[1]
-#         g_inverse_jacobian = reshape(g_inverse_jacobian,(4,4,4)) # reshape
-
-#         display(g_inverse_jacobian)
-#         # Second derivative of the metric 
-#         println("hessian")
-#         g_hessian = Zygote.hessian(x -> RelativisticDynamics.covariant_metric_zygote(x,a), [t r θ ϕ])[1]
-#         g_hessian = reshape(g_hessian,(4,4,4,4)) # reshape
-
-#         display(g_hessian)
-
[…]
