### 2023-06-14 #4992 @langfzac
After merging [tomkimpson/RelativisticDynamics.jl#42](https://github.com/tomkimpson/RelativisticDynamics.jl/pull/42#issuecomment-1588472777), all of my comments have been addressed.
