+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"functions_index/#Index-of-all-functions","page":"Functions Index","title":"Index of all functions","text":"","category":"section"},{"location":"functions_index/","page":"Functions Index","title":"Functions Index","text":"","category":"page"},{"location":"functions_index/","page":"Functions Index","title":"Functions Index","text":"Modules = [RelativisticDynamics]","category":"page"},{"location":"functions_index/#RelativisticDynamics.Constants","page":"Functions Index","title":"RelativisticDynamics.Constants","text":"C = Constants(P)\n\nA struct to hold all variables which are constant over the course of the integration. These are derived from the user-defined parameters \n\n\n\n\n\n","category":"type"},{"location":"functions_index/#RelativisticDynamics.Constants-Tuple{RelativisticDynamics.SystemParameters}","page":"Functions Index","title":"RelativisticDynamics.Constants","text":"Generator function for a Constants struct.\n\n\n\n\n\n","category":"method"},{"location":"functions_index/#RelativisticDynamics.Model","page":"Functions Index","title":"RelativisticDynamics.Model","text":"M = Model(P,C)\n\nThe model struct which holds all the parameters (P) and constants (C)\n\n\n\n\n\n","category":"type"},{"location":"functions_index/#RelativisticDynamics.PrognosticVariables","page":"Functions Index","title":"RelativisticDynamics.PrognosticVariables","text":"Struct holding the so-called 'prognostic' variables. 'Prognostic' terminology is borrowed from cliamte science where it refers to  any variables that are predicted via integration\n\n\n\n\n\n","category":"type"},{"location":"functions_index/#RelativisticDynamics.SystemParameters","page":"Functions Index","title":"RelativisticDynamics.SystemParameters","text":"P = Parameters(kwargs...)\n\nA struct to hold all model parameters that may be changed by the user. The struct uses keywords such that default values can be changed at creation. The default values of the keywords define the default model setup.\n\n\n\n\n\n","category":"type"},{"location":"functions_index/#RelativisticDynamics.ELQ-NTuple{5, Any}","page":"Functions Index","title":"RelativisticDynamics.ELQ","text":"E,L,Q = ELQ(a,α,e,ι,D)\n\nCalculate the energy, angular momentum and Carter constant given the Keplerian orbital parameters, and the BH spin/direction Specific to the Kerr metric, see Schmidt 2002 arXiv:0202090\n\n\n\n\n\n","category":"method"},{"location":"functions_index/#RelativisticDynamics.PlotTrajectory","page":"Functions Index","title":"RelativisticDynamics.PlotTrajectory","text":"PlotTrajectory(solution,model,dimensions=[1,2,3],savepath=\"\")\n\nPlot trajectory of a body. Assumes coordinates are Boyer Lindquist.  Plots in either 2D or 3D depending on specification of dimensions. Saves a low resolution PNG figure to disk in example_media/\n\n\n\n\n\n","category":"function"},{"location":"functions_index/#RelativisticDynamics.StackedPlot","page":"Functions Index","title":"RelativisticDynamics.StackedPlot","text":"StackedPlot(solution,model,savepath=\"\")\n\nPlot the x-y and x-z trajectory of a body on two separate subplots\n\n\n\n\n\n","category":"function"},{"location":"functions_index/#RelativisticDynamics.bounds_checks-Tuple{RelativisticDynamics.SystemParameters}","page":"Functions Index","title":"RelativisticDynamics.bounds_checks","text":"bounds_checks(P)\n\nLook before you leap - are the user-specified kwargs in parameters physical and reasonable?\n\n\n\n\n\n","c
[…]
