### 2023-06-15 #4992 @tomkimpson
https://github.com/tomkimpson/RelativisticDynamics.jl/pull/42 is now merged.

All comments from both reviewers have now been addressed.

The only comment that has not is @duetosymmetry recommendation to change the package name. Having thought about it, I think `RelativisticDynamics.jl` is ok since the context around application to MPD is clear from the documentation and the JOSS paper. The subtitle is _"Relativistic Spin-Orbital Dynamics in Julia"_ which I think also makes the application explicit.

If the reviewers disagree, I am happy to take to other naming suggestions - I am struggling to think of something succinct that follows the Julia [package naming guidelines](https://pkgdocs.julialang.org/dev/creating-packages/#Package-naming-guidelines). Perhaps just `RelativisticSpinOrbitalDynamics.jl`?

If we can agree on an appropriate package name here I will then implement the change.

Thanks all.
