### 2023-06-22 #4992 @dfm
Thanks for the ping @tomkimpson! And thanks @langfzac for your constructive and efficient review!!

@duetosymmetry — Thanks for your review and feedback so far! It looks like @tomkimpson has made significant changes in response to your comments (https://github.com/tomkimpson/RelativisticDynamics.jl/pull/41), and has [a remaining question](https://github.com/openjournals/joss-reviews/issues/4992#issuecomment-1592157848). Please take a look as soon as you have a chance to see if you have more suggestions and if you can check off anymore checklist items. Thanks!!
