### 2023-10-01 #4992 @arfon
@dfm – these are the most similar historical papers:

[DynamicalBilliards.jl: An easy-to-use, modular and extendable Julia package for Dynamical Billiard systems in two dimensions.](https://github.com/openjournals/joss-reviews/issues/458)
Submitting author: ```@Datseris```
Handling editor: ```@kyleniemeyer (Active)```
Reviewers: ```@ahwillia```
Similarity score: 0.8195286949018195

[DynamicalSystems.jl: A Julia software library for chaos and nonlinear dynamics](https://github.com/openjournals/joss-reviews/issues/598)
Submitting author: ```@Datseris```
Handling editor: ```@kyleniemeyer (Active)```
Reviewers: ```@dhhagan```
Similarity score: 0.8132166880827265

[NuclearToolkit.jl: A Julia package for nuclear structure calculations](https://github.com/openjournals/joss-reviews/issues/4694)
Submitting author: ```@SotaYoshida```
Handling editor: ```@rkurchin (Active)```
Reviewers: ```@mdavezac, @villaa```
Similarity score: 0.8131710416069782

[`QuasinormalModes.jl`: A Julia package for computing discrete eigenvalues of second order ODEs](https://github.com/openjournals/joss-reviews/issues/4077)
Submitting author: ```@lucass-carneiro```
Handling editor: ```@pdebuyl (Active)```
Reviewers: ```@JamieBamber, @cescalara```
Similarity score: 0.810427232362389

[iharm3D: Vectorized General Relativistic Magnetohydrodynamics](https://github.com/openjournals/joss-reviews/issues/3336)
Submitting author: ```@bprather```
Handling editor: ```@eloisabentivegna (Active)```
Reviewers: ```@bgiacoma, @cpalenzuela```
Similarity score: 0.8103136564020536
