### 2023-10-03 #4992 @dfm
Many thanks @tamasgal for agreeing to step in and pick up as a reviewer!! Thanks also to @tomkimpson for your patience with this process.

@tamasgal — You can comment `@editorialbot generate my checklist` on this thread to get your checklist and go from there. You'll find more info in the first comment on this thread, and [this comment](https://github.com/openjournals/joss-reviews/issues/4992#issuecomment-1336470364) that I left near the start. Please let me know if you have any questions as we go!
