### 2023-10-03 tomkimpson/RelativisticDynamics.jl#56 @tamasgal
Version constraints on dependencies (compat entry)
I suggest to add some version constraints in the `compat` section of the `Project.toml` file to ensure that people are pulling the correct dependencies and use a compatible Julia version.

After installing version 0.1.0 of `RelativisticDynamics.jl`, the following dependencies were resolved in a Julia 1.9.3 session on an M1 mac:

```
  [861a8166] Combinatorics v1.0.2
  [0c46a032] DifferentialEquations v7.11.0
  [b964fa9f] LaTeXStrings v1.3.0
  [d96e819e] Parameters v0.12.3
  [3cdcf5f2] RecipesBase v1.3.4
  [1ed8b502] SciMLSensitivity v7.40.0
  [bc48ee85] Tullio v0.3.5
  [e88e6eb3] Zygote v0.6.65
  [37e2e46d] LinearAlgebra
```

I also ran an installation using the official Julia 1.6.7 Docker image (I had to `apt install unzip` in the container since it was needed for some of the dependencies) and the only difference is `DifferentialEquations v7.10.0` (instead of 7.11.0, probably due to some upstream incompatibilities with Julia 1.6.7).

I assume that the following entry will work fine, but please amend ;)

```
[compat]
Combinatorics = "1"
DifferentialEquations = "7"
LaTeXStrings = "^1.3"
Parameters = "^0.12"
RecipesBase = "^1.3"
SciMLSensitivity = "7"
Tullio = "^0.3"
Zygote = "^0.6"
julia = "^1.6"  # according to the CI
```
