### 2023-10-03 #4992 @tamasgal
The package name `RelativisticDynamics` is fine in my opinion, there is no need to waste time on that ;) I was able to install the package without any issues and ran the test suite without any failing tests. Both notebooks (the demo and the one to reproduce the JOSS paper figures) executed fine in Jupyter and reproduced the results.

I opened a tiny issue regarding the compat specifications for dependencies and Julia itself: https://github.com/tomkimpson/RelativisticDynamics.jl/issues/56

Furthermore, I also suggest to add a simple `CONTRIBUTING.md` to the repository.

I am now waiting for the latest PDF to be generated.
