### 2023-10-03 #4992 @tamasgal
The paper is well-written and concise and the software package is a nice addition to JOSS, filling a niche gap in the field of relativistic spin-orbital dynamics which gained a lot of attention in the past years. Other than the two points above, I don't have any objections.
