+++ docs/src/contributing.md
@@ -1,0 +1,13 @@
+# Contributing
+
+We welcome any contributions to this package. 
+
+Some recommended areas are:
+
+* Taking the [MPD Equations](MPD.md) beyond second order in the expansion.
+* Extension to general spin supplementary conditions, beyond. 
+* Generalisation to arbitrary spacetime metrics, not just Kerr.
+* Calculation of gravitational waveforms using e.g. [numerical kludge](https://arxiv.org/abs/gr-qc/0607007) approach (or other).
+* Extension to hyperbolic systems, e.g. [Li et. al. 2021](https://arxiv.org/abs/2110.03494).
+
+Contributions in other areas are also more than welcome - please just [open a pull request](https://github.com/tomkimpson/RelativisticDynamics.jl/pulls) .
