### 2023-10-04 #4992 @tomkimpson
Hey @tamasgal , thanks for the review, it's very appreciated! I have addressed all your comments in the above PR.

I'll merge if all looks good to you?
