### 2023-10-05 #4992 @dfm
@tomkimpson — I've opened a small PR with some minor edits to the manuscript, please take a look and merge or let me know what you think.

Once you've done that:

1. Take one last read through the manuscript to make sure that you're happy with it (it's harder to make changes later!), especially the author name and affiliation. I've taken a pass and it looks good to me!
