+++ README.md
@@ -4,6 +4,8 @@
 [![Dev](https://img.shields.io/badge/docs-dev-blue.svg)](https://tomkimpson.github.io/RelativisticDynamics.jl/dev/)
 [![Build Status](https://github.com/tomkimpson/RelativisticDynamics.jl/actions/workflows/CI.yml/badge.svg?branch=main)](https://github.com/tomkimpson/RelativisticDynamics.jl/actions/workflows/CI.yml?query=branch%3Amain)
 [![codecov](https://codecov.io/gh/tomkimpson/RelativisticDynamics.jl/branch/main/graph/badge.svg?token=LpfCcTaxFQ)](https://codecov.io/gh/tomkimpson/RelativisticDynamics.jl)
+[![DOI](https://zenodo.org/badge/513167798.svg)](https://zenodo.org/badge/latestdoi/513167798)
+
 
 Welcome to RelativisticDynamics.jl!
 
