+++ versions.js
@@ -1,6 +1,7 @@
 var DOC_VERSIONS = [
   "stable",
   "v0.10.0",
+  "v0.1.2",
   "v0.1.1",
   "v0.1.0",
   "dev",
