### 2023-10-06 #4992 @tomkimpson
Thanks @dfm - I have merged your PR. To your other points:

1. I am happy with the manuscript
2. The version number is v0.1.2
3. An archived release has been created on [Zenodo](https://zenodo.org/record/8412240), doi:10.5281/zenodo.8412240
