### 2023-10-06 #4992 @dfm
Thanks! (As I mentioned above ☝️) Can you update the Zenodo metadata (title and authors) to exactly match the JOSS paper? You should have an "Edit" button at the top right of the record page if you're logged in.
