### 2023-10-07 tomkimpson/RelativisticDynamics.jl#56 @tomkimpson
Fixed by https://github.com/tomkimpson/RelativisticDynamics.jl/pull/57
