### 2023-10-07 #4992 @tomkimpson
> Thanks! (As I mentioned above ☝️) Can you update the Zenodo metadata (title and authors) to exactly match the JOSS paper? You should have an "Edit" button at the top right of the record page if you're logged in.

@dfm This is now done 👍
